Remove unnecessary call to retrieve cookie
Prior to the change, the userEmail variable to retrieve the authentication token, and the authorization to conduct the GET request would also require retrieval of the cookie from the local storage. However, with the change, the unnecessary retrieval process would be removed since useAuthState from context.js has the user's email and token credentials saved.