Remove imageUrl state and setImageUrl
Removed the use of the imageUrl state and the setImageUrl function to update the imageUrl state. The profile picture URL is now provided by the currentUser object which is a prop value from the AuthProvider provider. The image URL is updated in the client side through the use of the updateProfilePic reducer function.