Update reducer.js
The values profile_pic and list_of_products can be retrieved from the local storage. The user object has been updated to include profile_pic and list_of_products. Two new cases have been added to AuthReducer, a case for updating the profile picture, and a case for updating the lists of products.