Fix products appearing in the wrong list
Prior to this fix, the user was able to see their added products in a different list. For example, if the user was adding products to their 'Tools' list, they were able to see the added products in the 'Toys' list. This fix was done by having the products reference the respective list when being added.