Add otherUsers prop and conditional functionalities
An otherUser prop is used and passed down to several components. The purpose of the otherUser prop is to check if the user is visiting other pages. If the user is currently on their own user dashboard, the otherUser prop will not exist in the other components. However, if the user is visiting other pages, the otherUser prop will be an object that contains the data of the other user's profile.