Fix bug when returning to previous page
If the user clicks on the Deals Mate image from another user's page, it would return the user back to the user's dashboard. However, if the user were to click on the back button, it would not return the user back to the previous user's page that they were just on. Instead, the user would see the json of the user that they were just viewing. With this fix, the user can now return back to the other user's page.