Update AddListDialogue
Updated the dialogue for adding a new list to allow the user to set their list's privacy setting. The privacy setting is defaulted to being public.

Several bug fixes were also made. Prior to the update, the user was able to make multiple lists with the same list title. However, this would cause a bug and would not allow the user to visit the list's content. The update denies the user to successfully create a list by checking if the list title they added was the same as any existing lists' title. In addition, the imageUrl parameter for the addList function is set to be an empty string as a default value.