Map the current user's list of friends and parse each of the JSON string of friends
Prior to this change, the list of friends only contained the followed user's email. Since the list of friends now contain JSON strings of followed users, we must parse each JSON string back into a proper JSON object. With the new parsedListOfFriends, the .some() function will look to see if the current user is following any users with the same email as the other user's profile.