Move getFriends() function from action.js to here and changed fetch method from PUT to GET
The getFriends() function is moved from the actions.js file to here. In addition, the getFriends() function method was changed from using a PUT method to fetch data to using a GET method. Since the fetched data of followed users is an array of JSON strings, the friendDetails array was mapped twice. The first map is done to parse the JSON string back to a JSON object and the second map is to map the FriendCard component with the proper friend details.